test(rbac): improve docs and dev commands
