+++ workspaces/rbac/package.json
@@ -6,6 +6,9 @@
     "node": "22 || 24"
   },
   "scripts": {
+    "start": "backstage-cli repo start",
+    "start:multi-user": "yarn workspaces foreach -A --include @backstage-community/plugin-rbac --include @backstage-community/plugin-rbac-backend --parallel -v -i run start:multi-user",
+    "start:legacy": "yarn workspaces foreach -A --include @backstage-community/plugin-rbac --include @backstage-community/plugin-rbac-backend --parallel -v -i run start:legacy",
     "tsc": "tsc",
     "tsc:full": "tsc --skipLibCheck true --incremental false",
     "build:all": "backstage-cli repo build --all",
